MC-27: Move 'lib' code into 'modules' directory
Inspired by https://hackernoon.com/fractal-a-react-app-structure-for-infinite-scale-4dab943092af.
